Stop notebook from executing when any widget is changed
When a widget is changed in a notebook that notebook will automatically execute the cell that grabs the value.  Since the function that grabs the value does all the other steps to send the request as well, when any widget value was changed it would send a request.
Separated the step that grabs the widget values into a separate cell so that a request isn't sent until the customer executes notebook.